gpu: nvgpu: fb: fix MISRA 10.3 violations
This fixes a number of MISRA Rule 10.3 violations for assignment of
objects of different essential type or size in the common/fb unit.

JIRA NVGPU-2954

Reviewed-on: https://git-master.nvidia.com/r/2075596
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/hal/fb/fb_gp10b.c
@@ -34,5 +34,5 @@ u64 gp10b_fb_compression_page_size(struct gk20a *g)
 
 unsigned int gp10b_fb_compressible_page_size(struct gk20a *g)
 {
-	return SZ_4K;
+	return (unsigned int)SZ_4K;
 }
